python_service/requirements.txt
fastapi
uvicorn
google-adk>=1.0.0
google-genai>=1.9.0
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
pydantic>=2.10.6
python-dotenv>=1.0.1
python-multipart
firecrawl-py>=4.5.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
firebase-admin>=6.5.0
duckduckgo-search>=6.0.0